+++ src/design/plone/contenttypes/upgrades/to_730x.py
@@ -11,6 +11,18 @@ logger = logging.getLogger(__name__)
 DEFAULT_PROFILE = "profile-design.plone.contenttypes:default"
 
 
+def update_profile(context, profile, run_dependencies=True):
+    context.runImportStepFromProfile(DEFAULT_PROFILE, profile, run_dependencies)
+
+
+def update_catalog(context):
+    update_profile(context, "catalog")
+
+
+def update_registry(context):
+    update_profile(context, "plone.app.registry", run_dependencies=False)
+
+
 def to_7301(context):
     brains = api.content.find(portal_type="Persona")
     for brain in brains:
@@ -43,3 +55,21 @@ def to_7301(context):
             if api.content.get_state(child) != "published":
                 with api.env.adopt_roles(["Reviewer"]):
                     api.content.transition(obj=child, transition="publish")
+
+
+def to_7302(context):
+    update_catalog(context)
+    brains = api.content.find(portal_type="Event")
+    tot = len(brains)
+    i = 0
+    for brain in brains:
+        i += 1
+        if i % 100 == 0:
+            logger.info("Progress: {}/{}".format(i, tot))
+        event = brain.getObject()
+        event.reindexObject(idxs=["rassegna"])
+
+
+def to_7303(context):
+    update_registry(context)
+    logger.info("Update registry")